qty is integer, not float

+++ src/resources/js/rsvp-block.js
@@ -194,7 +194,7 @@ var tribe_tickets_rsvp_block = {
 		return (
 			$.trim( $name.val() ).length &&
 				$.trim( $email.val() ).length &&
-				parseFloat( $qty.val() ) > 0
+				parseInt( $qty.val() ) > 0
 		);
 	};
